fixed cmake; fixed dockerfile (install specific version of gcc); fixed bug in parsing timestamps; fixed block hashing bug (from timestamp parsing bug); replaced SortedDigestSerializer with ToJSONSerializer; added environment check to Miner startup; bumped submodules

+++ vol-node-cpp/src/volition/AccountODBM.h
@@ -25,7 +25,7 @@ public:
     Account::Index          mIndex;
 
     LedgerFieldODBM < string >      mBody;
-    LedgerFieldODBM < size_t >      mAssetCount;
+    LedgerFieldODBM < u64 >         mAssetCount;
     LedgerFieldODBM < u64 >         mInventoryNonce;
     LedgerFieldODBM < string >      mMinerInfo;
     LedgerFieldODBM < u64 >         mTransactionNonce;
